py-utils/python_requirements.txt
aiohttp==3.8.1
configparser==4.0.2
confluent-kafka==1.5.0
cryptography==3.3.2
dict2xml==1.6.1
elasticsearch-dsl==7.3.0
elasticsearch==7.12.0
matplotlib==3.1.3
networkx==2.4
paramiko==2.10.4
prettytable==0.7.2
python-consul==1.1.0
python-crontab==2.5.1
PyYAML==5.4.1
schematics==2.1.0
toml==0.10.0
